// airbag/frontend/src/types/response/response.quotation.ts
import { TypeColorAllResponse } from "./response.color";
import { quotationRepair } from "./response.quotation_repair";
import { Type_MS_BRAND } from "./response.ms-brand";
import { Type_MS_BRANDMODEL_All } from "./response.ms-brandmodel";
import { Type_MS_CompaniesL_All } from "@/types/response/response.ms-companies";

export type QUOTATION_ALL = {
  quotation_id: string;
  quotation_doc: string;
  quotation_date: string | null;
  customer_id: string;
  customer_name: string | null;
  contact_number: string | null;
  line_id: string | null;
  position: string | null;
  addr_map: string;
  addr_number: string;
  addr_alley: string;
  addr_street: string;
  addr_subdistrict: string;
  addr_district: string;
  addr_province: string;
  addr_postcode: string;
  image_url: string | null;
  brand_id: string | null;
  model_id: string | null;
  car_year: string | null;
  car_color_id: string | null;
  total_price: string | null;
  tax: number | null;
  deadline_day: number | null;
  appointment_date: string | null;
  remark: string | null;
  insurance: boolean;
  insurance_date: string;
  repair_summary: string | null;
  is_box_detail: boolean;
  lock: boolean;
  quotation_status: string | null;
  responsible_by: string;
  responsible_date: string;
  approval_date: string | null;
  approval_by: string | null;
  approval_notes: string | null;
  // deal_closed_status: boolean;
  deal_closed_date: string | null;
  deal_closed_by: string | null;
  created_at: string;
  created_by: string;
  updated_at: string;
  updated_by: string;
  is_deleted: boolean;
  master_customer: master_customer;
  master_brand: Type_MS_BRAND;
  quotationRepair: quotationRepair[];
  master_brandmodel: Type_MS_BRANDMODEL_All;
  master_color: TypeColorAllResponse;
  companies: Type_MS_CompaniesL_All;
};

type master_customer = {
  customer_id: string;
  company_id: string;
  customer_tin: string | null;
  customer_code: string;
  customer_prefix: string | null;
  customer_name: string | null;
  contact_name: string | null;
  customer_position: string | null;
  contact_number: string | null;
  line_id: string | null;
  addr_number: string | null;
  addr_alley: string | null;
  addr_street: string | null;
  addr_subdistrict: string | null;
  addr_district: string | null;
  addr_province: string | null;
  addr_postcode: string | null;
  payment_terms: string | null;
  payment_terms_day: string | null;
  tax: string | null;
  comment_customer: string | null;
  comment_sale: string | null;
  competitor: string | null;
  created_at: string;
  created_by: string;
  updated_by: string;
  updated_at: string | null;
};

export type QuotationResponse = {
  success: boolean;
  message: string;
  responseObject: QUOTATION_ALL;
  statusCode: number;
};